+++ include/libpkgstate-apply/libpkgstate-apply.h
@@ -2,5 +2,7 @@
 // SPDX-License-Identifier: GPL-3.0-or-later
 #pragma once
 
+#include <libpkgstate-apply/export.h>
+
 #include <libpkgstate-apply/adapter.h>
 #include <libpkgstate-apply/state_projection.h>